# docs/swagger.yaml
definitions:
...
  models.BrandCar:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Car:
    properties:
      brand:
        $ref: '#/definitions/models.BrandCar'
      brand_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_second:
        type: boolean
      name:
        type: string
      price:
        type: number
      type:
        $ref: '#/definitions/models.TypeCar'
      type_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.TypeCar:
    properties:
      cars:
        items:
          $ref: '#/definitions/models.Car'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/cms/type-cars:
    get:
      description: Get a list of all type cars
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TypeCar'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Get all type cars
      tags:
      - type-cars
    post:
      consumes:
      - application/json
      description: Create a new type car
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Type Car object
        in: body
        name: type_car
        required: true
        schema:
          $ref: '#/definitions/models.TypeCar'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TypeCar'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Create a new type car
      tags:
      - type-cars
  /api/cms/type-cars/{id}:
    delete:
      description: Delete a specific type car
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Type Car ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Delete a type car
      tags:
      - type-cars
    get:
      description: Get details of a specific type car including associated cars
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Type Car ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TypeCar'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Get a type car by ID
      tags:
      - type-cars
    put:
      consumes:
      - application/json
      description: Update details of a specific type car
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Type Car ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Type Car object
        in: body
        name: type_car
        required: true
        schema:
          $ref: '#/definitions/models.TypeCar'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TypeCar'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Update a type car
      tags:
      - type-cars
...
swagger: "2.0"